Refactor: V-1.7.1 Restructure app architecture to fit app architecture
- Cleaned up file organization without changing functionality.
- Incremented PATCH version to 1.7.1 for housekeeping changes.
