+++ .idea/deploymentTargetSelector.xml
@@ -4,10 +4,10 @@
     <selectionStates>
       <SelectionState runConfigName="app">
         <option name="selectionMode" value="DROPDOWN" />
-        <DropdownSelection timestamp="2024-11-20T13:38:53.521831500Z">
+        <DropdownSelection timestamp="2024-11-23T14:05:50.267720200Z">
           <Target type="DEFAULT_BOOT">
             <handle>
-              <DeviceId pluginId="PhysicalDevice" identifier="serial=R58T32DVTMP" />
+              <DeviceId pluginId="PhysicalDevice" identifier="serial=cd7751cf" />
             </handle>
           </Target>
         </DropdownSelection>
