+++ app/build.gradle.kts
@@ -7,7 +7,7 @@ plugins {
 android {
     val versionMajor = 1
     val versionMinor = 7
-    val versionPatch = 0
+    val versionPatch = 1
     val appName = "KhoroojYar"
     val appVersionCode = versionMajor * 10000 + versionMinor * 100 + versionPatch
     val appVersionName = "$versionMajor.$versionMinor.$versionPatch"
