+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.KhoroojYar"
         tools:targetApi="31">
         <activity
-            android:name="com.faridnia.khoroojyar.MainActivity"
+            android:name="com.faridnia.khoroojyar.presentation.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.KhoroojYar">
