+++ app/src/main/java/com/faridnia/khoroojyar/presentation/component/snackbar/SnackbarAction.kt
@@ -1,4 +1,4 @@
-package com.faridnia.khoroojyar.ui.component.snackbar
+package com.faridnia.khoroojyar.presentation.component.snackbar
 
 data class SnackbarAction(
     val name: String,
