+++ app/src/main/java/com/faridnia/khoroojyar/presentation/component/snackbar/SnackbarType.kt
@@ -1,4 +1,4 @@
-package com.faridnia.khoroojyar.ui.component.snackbar
+package com.faridnia.khoroojyar.presentation.component.snackbar
 
 import androidx.compose.ui.graphics.Color
