+++ app/src/main/java/com/faridnia/khoroojyar/presentation/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.faridnia.khoroojyar.ui.theme
+package com.faridnia.khoroojyar.presentation.theme
 
 import androidx.compose.ui.graphics.Color
