+++ app/src/main/java/com/faridnia/khoroojyar/presentation/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.faridnia.khoroojyar.ui.theme
+package com.faridnia.khoroojyar.presentation.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
